+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - @aofl/cli - i18n ignores _r & _c keys when validating the manifest
 - @aofl/cli - excluded @webcomponents from webpack config ignore
 - @aofl/router - fixed matching dynamic routes without trailing slash
+- @aofl/web-components/aofl-select-list - Prevent dispatching change event while initializing
 
 ### [Security]